split entrypoint for vpn service
