+++ src/vpn_service.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python
+# Copyright (C) 2020 OpenMotics BV
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from __future__ import absolute_import
+
+from platform_utils import System
+System.import_libs()
+
+import logging
+
+from bus.om_bus_service import MessageService
+from gateway.initialize import initialize
+from gateway.services.vpn import VPNService
+
+logger = logging.getLogger('openmotics')
+
+
+def setup_logger():
+    """ Setup the OpenMotics logger. """
+
+    logger.setLevel(logging.INFO)
+    logger.propagate = False
+
+    handler = logging.StreamHandler()
+    handler.setLevel(logging.INFO)
+    handler.setFormatter(logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s'))
+    logger.addHandler(handler)
+
+
+if __name__ == '__main__':
+    setup_logger()
+    initialize(message_client_name='vpn_service')
+
+    logger.info('Starting VPN service')
+    vpn_service = VPNService()
+    vpn_service.start()
